feat: reworked options and injectables
BREAKING CHANGE: Structure and options
