+++ packages/ldap-auth-backend/src/index.ts
@@ -1,8 +1,4 @@
+export { prepareBackstageIdentityResponse } from './auth';
 export { ldap, ProviderLdapAuthProvider } from './provider';
-export {
-    JWTTokenValidator,
-    parseJwtPayload,
-    prepareBackstageIdentityResponse,
-    normalizeTime,
-} from './helpers';
-export type { TokenValidator } from './helpers';
+export { JWTTokenValidator, parseJwtPayload, normalizeTime } from './jwt';
+export type { TokenValidator } from './jwt';
